@@ -1,4 +1,4 @@
-name: gestures
+name: gestures_app
 description: A new app that shows how to detect gestures on a map view with the HERE SDK for Flutter.
 
 # The following line prevents the package from being accidentally published to
@@ -18,8 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
-  flutter: ">=1.22.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.6"
 
 dependencies:
   flutter:
