@@ -1,4 +1,4 @@
-name: mapitems
+name: map_items_app
 description: A new Flutter app that uses the HERE SDK for Flutter to pin images to the map.
 
 # The following line prevents the package from being accidentally published to
@@ -18,8 +18,8 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
-  flutter: ">=1.22.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.6"
 
 dependencies:
   flutter:
