@@ -34,16 +34,15 @@ typedef ShowDialogFunction = void Function(String title, String message);
 class SearchExample {
   HereMapController _hereMapController;
   MapCamera _camera;
-  MapImage _poiMapImage;
+  MapImage? _poiMapImage;
   List<MapMarker> _mapMarkerList = [];
-  SearchEngine _searchEngine;
+  late SearchEngine _searchEngine;
   ShowDialogFunction _showDialog;
 
-  SearchExample(Function showDialogCallback, HereMapController hereMapController) {
-    _showDialog = showDialogCallback;
-    _hereMapController = hereMapController;
-    _camera = _hereMapController.camera;
-
+  SearchExample(ShowDialogFunction showDialogCallback, HereMapController hereMapController)
+      : _showDialog = showDialogCallback,
+        _hereMapController = hereMapController,
+        _camera = hereMapController.camera {
     double distanceToEarthInMeters = 10000;
     _camera.lookAtPointWithDistance(GeoCoordinates(52.520798, 13.409408), distanceToEarthInMeters);
 
@@ -91,16 +90,15 @@ class SearchExample {
   }
 
   void _setTapGestureHandler() {
-    _hereMapController.gestures.tapListener = TapListener.fromLambdas(lambda_onTap: (Point2D touchPoint) {
+    _hereMapController.gestures.tapListener = TapListener((Point2D touchPoint) {
       _pickMapMarker(touchPoint);
     });
   }
 
   void _setLongPressGestureHandler() {
-    _hereMapController.gestures.longPressListener =
-        LongPressListener.fromLambdas(lambda_onLongPress: (GestureState gestureState, Point2D touchPoint) {
+    _hereMapController.gestures.longPressListener = LongPressListener((GestureState gestureState, Point2D touchPoint) {
       if (gestureState == GestureState.begin) {
-        GeoCoordinates geoCoordinates = _hereMapController.viewToGeoCoordinates(touchPoint);
+        GeoCoordinates? geoCoordinates = _hereMapController.viewToGeoCoordinates(touchPoint);
         if (geoCoordinates == null) {
           return;
         }
@@ -115,20 +113,24 @@ class SearchExample {
     SearchOptions reverseGeocodingOptions = SearchOptions(LanguageCode.enGb, maxItems);
 
     _searchEngine.searchByCoordinates(geoCoordinates, reverseGeocodingOptions,
-        (SearchError searchError, List<Place> list) async {
+        (SearchError? searchError, List<Place>? list) async {
       if (searchError != null) {
         _showDialog("Reverse geocoding", "Error: " + searchError.toString());
         return;
       }
 
       // If error is null, list is guaranteed to be not empty.
-      _showDialog("Reverse geocoded address:", list.first.address.addressText);
+      _showDialog("Reverse geocoded address:", list!.first.address.addressText);
     });
   }
 
   void _pickMapMarker(Point2D touchPoint) {
     double radiusInPixel = 2;
     _hereMapController.pickMapItems(touchPoint, radiusInPixel, (pickMapItemsResult) {
+      if (pickMapItemsResult == null) {
+        // Pick operation failed.
+        return;
+      }
       List<MapMarker> mapMarkerList = pickMapItemsResult.markers;
       if (mapMarkerList.length == 0) {
         print("No map markers found.");
@@ -136,9 +138,9 @@ class SearchExample {
       }
 
       MapMarker topmostMapMarker = mapMarkerList.first;
-      Metadata metadata = topmostMapMarker.metadata;
+      Metadata? metadata = topmostMapMarker.metadata;
       if (metadata != null) {
-        CustomMetadataValue customMetadataValue = metadata.getCustomValue("key_search_result");
+        CustomMetadataValue? customMetadataValue = metadata.getCustomValue("key_search_result");
         if (customMetadataValue != null) {
           SearchResultMetadata searchResultMetadata = customMetadataValue as SearchResultMetadata;
           String title = searchResultMetadata.searchResult.title;
@@ -163,14 +165,14 @@ class SearchExample {
     int maxItems = 30;
     SearchOptions searchOptions = SearchOptions(LanguageCode.enUs, maxItems);
 
-    _searchEngine.searchByText(query, searchOptions, (SearchError searchError, List<Place> list) async {
+    _searchEngine.searchByText(query, searchOptions, (SearchError? searchError, List<Place>? list) async {
       if (searchError != null) {
         _showDialog("Search", "Error: " + searchError.toString());
         return;
       }
 
       // If error is null, list is guaranteed to be not empty.
-      int listLength = list.length;
+      int listLength = list!.length;
       _showDialog("Search for $queryString", "Results: $listLength. Tap marker to see details.");
 
       // Add new marker for each search result on map.
@@ -178,7 +180,7 @@ class SearchExample {
         Metadata metadata = Metadata();
         metadata.setCustomValue("key_search_result", SearchResultMetadata(searchResult));
         // Note: getGeoCoordinates() may return null only for Suggestions.
-        addPoiMapMarker(searchResult.geoCoordinates, metadata);
+        addPoiMapMarker(searchResult.geoCoordinates!, metadata);
       }
     });
   }
@@ -193,7 +195,7 @@ class SearchExample {
         TextQuery.withAreaCenter(
             "p", // User typed "p".
             centerGeoCoordinates),
-        searchOptions, (SearchError searchError, List<Suggestion> list) async {
+        searchOptions, (SearchError? searchError, List<Suggestion>? list) async {
       _handleSuggestionResults(searchError, list);
     });
 
@@ -201,7 +203,7 @@ class SearchExample {
         TextQuery.withAreaCenter(
             "pi", // User typed "pi".
             centerGeoCoordinates),
-        searchOptions, (SearchError searchError, List<Suggestion> list) async {
+        searchOptions, (SearchError? searchError, List<Suggestion>? list) async {
       _handleSuggestionResults(searchError, list);
     });
 
@@ -209,24 +211,24 @@ class SearchExample {
         TextQuery.withAreaCenter(
             "piz", // User typed "piz".
             centerGeoCoordinates),
-        searchOptions, (SearchError searchError, List<Suggestion> list) async {
+        searchOptions, (SearchError? searchError, List<Suggestion>? list) async {
       _handleSuggestionResults(searchError, list);
     });
   }
 
-  void _handleSuggestionResults(SearchError searchError, List<Suggestion> list) {
+  void _handleSuggestionResults(SearchError? searchError, List<Suggestion>? list) {
     if (searchError != null) {
       print("Autosuggest Error: " + searchError.toString());
       return;
     }
 
     // If error is null, list is guaranteed to be not empty.
-    int listLength = list.length;
+    int listLength = list!.length;
     print("Autosuggest results: $listLength.");
 
     for (Suggestion autosuggestResult in list) {
       String addressText = "Not a place.";
-      Place place = autosuggestResult.place;
+      Place? place = autosuggestResult.place;
       if (place != null) {
         addressText = place.address.addressText;
       }
@@ -243,7 +245,7 @@ class SearchExample {
     int maxItems = 30;
     SearchOptions searchOptions = SearchOptions(LanguageCode.deDe, maxItems);
 
-    _searchEngine.searchByAddress(query, searchOptions, (SearchError searchError, List<Place> list) async {
+    _searchEngine.searchByAddress(query, searchOptions, (SearchError? searchError, List<Place>? list) async {
       if (searchError != null) {
         _showDialog("Geocoding", "Error: " + searchError.toString());
         return;
@@ -251,9 +253,10 @@ class SearchExample {
 
       String locationDetails = "";
 
-      for (Place geocodingResult in list) {
+      // If error is null, list is guaranteed to be not empty.
+      for (Place geocodingResult in list!) {
         // Note: getGeoCoordinates() may return null only for Suggestions.
-        GeoCoordinates geoCoordinates = geocodingResult.geoCoordinates;
+        GeoCoordinates geoCoordinates = geocodingResult.geoCoordinates!;
         Address address = geocodingResult.address;
         locationDetails = address.addressText +
             ". GeoCoordinates: " +
@@ -277,7 +280,7 @@ class SearchExample {
       _poiMapImage = MapImage.withPixelDataAndImageFormat(imagePixelData, ImageFormat.png);
     }
 
-    MapMarker mapMarker = MapMarker(geoCoordinates, _poiMapImage);
+    MapMarker mapMarker = MapMarker(geoCoordinates, _poiMapImage!);
     _hereMapController.mapScene.addMapMarker(mapMarker);
     _mapMarkerList.add(mapMarker);
 
@@ -300,9 +303,14 @@ class SearchExample {
   }
 
   GeoBox _getMapViewGeoBox() {
-    GeoBox geoBox = _camera.boundingBox;
+    GeoBox? geoBox = _camera.boundingBox;
     if (geoBox == null) {
-      throw Exception("GeoBox creation failed, corners are null.");
+      print("GeoBox creation failed, corners are null. This can happen when the map is tilted. Falling back to a fixed box.");
+      GeoCoordinates southWestCorner = GeoCoordinates(
+          _camera.state.targetCoordinates.latitude - 0.05, _camera.state.targetCoordinates.longitude - 0.05);
+      GeoCoordinates northEastCorner = GeoCoordinates(
+          _camera.state.targetCoordinates.latitude + 0.05, _camera.state.targetCoordinates.longitude + 0.05);
+      geoBox = GeoBox(southWestCorner, northEastCorner);
     }
     return geoBox;
   }
