@@ -35,7 +35,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  RoutingExample _routingExample;
+  RoutingExample? _routingExample;
 
   @override
   Widget build(BuildContext context) {
@@ -72,7 +72,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _onMapCreated(HereMapController hereMapController) {
-    hereMapController.mapScene.loadSceneForMapScheme(MapScheme.normalDay, (MapError error) {
+    hereMapController.mapScene.loadSceneForMapScheme(MapScheme.normalDay, (MapError? error) {
       if (error == null) {
         _routingExample = RoutingExample(_showDialog, hereMapController);
       } else {
@@ -82,23 +82,23 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _addRouteButtonClicked() {
-    _routingExample.addRoute();
+    _routingExample?.addRoute();
   }
 
   void _addWaypointsButtonClicked() {
-    _routingExample.addWaypoints();
+    _routingExample?.addWaypoints();
   }
 
   void _onlineButtonClicked() {
-    _routingExample.useOnlineRoutingEngine();
+    _routingExample?.useOnlineRoutingEngine();
   }
 
   void _offlineButtonClicked() {
-    _routingExample.useOfflineRoutingEngine();
+    _routingExample?.useOfflineRoutingEngine();
   }
 
   void _clearMapButtonClicked() {
-    _routingExample.clearMap();
+    _routingExample?.clearMap();
   }
 
   // A helper method to add a button on top of the HERE map.
